% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars Data Frame}
\format{A data frame with 234 rows and 3 variables:
\describe{
  \item{\code{hwy}}{integer highway miles per gallon}
  \item{\code{class}}{character "type" of car}
  \item{\code{year}}{integer year}
}}
\description{
This is a subset of the mpg dataframe distributed with the \code{ggplot2} package
}
\keyword{cars}
