% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Coagualtaion_data.R
\docType{data}
\name{coagulation}
\alias{coagulation}
\title{Coagualtion data from library of faraway}
\format{A data frame with 24 rows and two columns containing information:

\describe{
   \item{coag}: coagulation time in seconds
   \item{diet}: diet type - A,B,C or D
}}
\source{
"Statistics for Experimenters" by G. P. Box, W. G. Hunter and J. S. Hunter, Wiley, 1978
}
\usage{
coagulation
}
\description{
Coagualtion data from library of faraway
}
\keyword{datasets}

