% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{lsmeans.oneway}
\alias{lsmeans.oneway}
\title{lsmeans.oneway gets the input from the oneway.defual to calculate the LSD for it}
\usage{
lsmeans.oneway(object, ...)
}
\arguments{
\item{object}{object is the ouput of oneway.default which has the values of anova table}

\item{...}{gets an additional argument}
}
\description{
lsmeans.oneway gets the input from the oneway.defual to calculate the LSD for it
}

