% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway.default}
\alias{oneway.default}
\title{oneway.defualt for calculating the anova parameters}
\usage{
\method{oneway}{default}(z, ...)
}
\arguments{
\item{z}{z is the input in form of list/data.frame}

\item{...}{... is the addistion arguments here}
}
\description{
oneway.defualt for calculating the anova parameters
}
\examples{
onewayanova:::oneway.default(coagulation)
}

