% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway.factor}
\alias{oneway.factor}
\title{oneway.factor combines two columns of factor and values togethere and then call the oneway.default function to do the calculations for anova}
\usage{
\method{oneway}{factor}(z, y, ...)
}
\arguments{
\item{z}{as a factor input here}

\item{y}{as values input here}
}
\description{
oneway.factor combines two columns of factor and values togethere and then call the oneway.default function to do the calculations for anova
}
\examples{
library(faraway)
z <- coagulation$diet
y <- coagulation$coa
aa<- onewayanova:::oneway.factor(z,y)
onewayanova:::print.oneway(aa)
}
\seealso{
\code{\link{print.oneway}}

\code{\link{print.summary.oneway}}
}

