% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway.formula}
\alias{oneway.formula}
\title{oneway.formula gets a data list and extracts two specific colmumns from the formula in order to make it ready to find the anova variables by passing it to oneway.default}
\usage{
\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{formula is regular formula in r which is defining two variables by ~ sign}

\item{data}{data is a list that can have more two columns so its two col will be extracted by formula}
}
\description{
oneway.formula gets a data list and extracts two specific colmumns from the formula in order to make it ready to find the anova variables by passing it to oneway.default
}

