% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{One Way Analysis of Variance}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{an object of class formula, relating the dependent variable to the grouping variable}

\item{data}{a dataframe containing the variables in the model}
}
\value{
a list with two elements:
\item{oneway}{a list with the lm results}
\item{summarystats}{a dataframe with the summary statistics}
}
\description{
\code{oneway} computes a one way analysis of variance and includes group-level summary statistics.
}
\details{
This function computes a standard oneway ANOVA, means, and standard deviation.
Missing values are handled via listwise deletion.
}
\examples{
mileage<-oneway(hwy ~ class, cars)
summary(mileage)
print(mileage)
plot(mileage)
}
\author{
Rob Kabacoff <rkabacoff@wesleyan.edu>
}
