% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayAnova-package.R
\docType{package}
\name{onewayAnova-package}
\alias{onewayAnova}
\alias{onewayAnova-package}
\title{onewayAnova: Performs a one way ANOVA}
\description{
This package is a wonderful example of having too much time on your hands.
}
\author{
\strong{Maintainer}: Kaitlyn Thomas-Franz \email{kthomasfranz@wesleyan.edu}

}
\keyword{internal}
