% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{print.oneway}
\alias{print.oneway}
\title{print.oneway gets input from the oneway.defual and print the very basic information about anova}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{x is the input from oneway.default which has the values of anova}

\item{...}{... is the addition argument here}
}
\description{
print.oneway gets input from the oneway.defual and print the very basic information about anova
}
\examples{
library(faraway)
data(coagulation)

attach(coagulation)
k <- onewayanova:::oneway.default(coagulation)
onewayanova:::print.oneway(k)
}

