% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{print.summary.oneway}
\alias{print.summary.oneway}
\title{print.summary.oneway gets the input from the summary.oneway to print it on the screen}
\usage{
\method{print}{summary.oneway}(x, ...)
}
\arguments{
\item{x}{x  is the ouput of summary.oneway which has the values of anova table}

\item{...}{... is an addition argument}
}
\description{
print.summary.oneway gets the input from the summary.oneway to print it on the screen
}
\examples{
library(faraway)
data(coagulation)
attach(coagulation)
oo <- onewayanova:::oneway.default(coagulation)
pp <- onewayanova:::summary.oneway(oo)
onewayanova:::print.summary.oneway(pp)
}

