% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Providing summary information for the output of oneway.defual}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{object is the ouput of oneway.default which has the values of anova table}

\item{...}{is an addition argument}
}
\description{
Providing summary information for the output of oneway.defual
}
\examples{
library(faraway)
data(coagulation)
attach(coagulation)
k <- onewayanova:::oneway.default(coagulation)
kk <- onewayanova:::summary.oneway(k)
onewayanova:::print.oneway(k)
}
\seealso{
\code{\link{print.summary.oneway}}
}

