% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.r
\docType{data}
\name{cars}
\alias{cars}
\title{cars dataset}
\format{A data frame with 234 rows and 3 variables:
\describe{
  \item{\code{hwy}}{integer COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{year}}{integer COLUMN_DESCRIPTION}
}}
\usage{
cars
}
\description{
Subset of mpg dataset from ggplot2 containing hwy class and year
}
\details{
DETAILS
}
\keyword{datasets}
