% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars Data Frame}
\format{A data frame with 234 rows and 3 variables:
\describe{
  \item{\code{hwy}}{integer Highway MPG}
  \item{\code{class}}{character Car Class}
  \item{\code{year}}{integer Year}
}}
\usage{
cars
}
\description{
This is a subset of the MPG data frame distributed with
the \code{ggplot2} package
}
\keyword{cars}
