% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway.default}
\alias{oneway.default}
\title{Calculating the anova parameters}
\usage{
\method{oneway}{default}(z, ...)
}
\arguments{
\item{z}{is the input in form of list or data.frame}

\item{...}{is the addition arguments here}
}
\description{
Calculating the anova parameters
}
\examples{
a <- onewayanova:::oneway.default(coagulation)
onewayanova:::print.oneway(a)
}
\seealso{
\code{\link{print.oneway}}

\code{\link{print.summary.oneway}}
}

