% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway}
\alias{oneway}
\title{oneway for implementing One-Way S3}
\usage{
oneway(z, ...)
}
\arguments{
\item{z}{z is the input in form of list/data.frame}

\item{...}{... is the addistion arguments here}
}
\description{
oneway for implementing One-Way S3
}

