% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{oneway}
\alias{oneway}
\title{Implementing One-Way S3 update}
\usage{
oneway(z, ...)
}
\arguments{
\item{z}{is the input in form of list OR data.frame}

\item{...}{is the addistion arguments here}
}
\description{
Implementing One-Way S3 update
}
\seealso{
\code{\link[aov]{stat}}

\code{\link{print.oneway}}

\code{\link{print.summary.oneway}}
}

