% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayAnova-package.R
\docType{package}
\name{onewayAnova-package}
\alias{onewayAnova}
\alias{onewayAnova-package}
\title{onewayAnova: Perfomrs a One Way ANOVA}
\description{
This package is a marvelous example of having 
  too much time on your hands.
}
\author{
\strong{Maintainer}: Jiner Zheng \email{first.last@example.com}

}
\keyword{internal}
