% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayAnova-package.R
\docType{package}
\name{onewayAnova-package}
\alias{onewayAnova}
\alias{onewayAnova-package}
\title{onewayAnova: Performs a One Way ANOVA}
\description{
This package is a marvelous example of having too much time on your
  hands.
}
\author{
\strong{Maintainer}: Nurah Jaradat \email{first.last@example.com}

}
\keyword{internal}
