% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{plot.oneway}
\alias{plot.oneway}
\title{plot.oneway gets input from the oneway.defualt to plot the boxes side by side to give the information about each group}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{x is the ouput of oneway.default which has the values of input list}

\item{...}{... is and additional argument}
}
\description{
plot.oneway gets input from the oneway.defualt to plot the boxes side by side to give the information about each group
}
\examples{
library(faraway)
data(coagulation)
attach(coagulation)
ll <- onewayanova:::oneway.default(coagulation)
onewayanova:::plot.oneway(ll)
}

