% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.oneway.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Boxplot for One Way ANOVA object}
\usage{
\method{plot}{oneway}(x, ...)
}
\arguments{
\item{x}{object of class \code{oneway}}

\item{...}{parameters passed to the boxplot function}
}
\description{
creates a boxplot for an oneway object
}
\examples{
mileage <- oneway(hwy ~ class, cars)
plot(mileage)
}
