% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{print.oneway}
\alias{print.oneway}
\title{Printing DF and SS}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{x is the input from oneway.default which has the values of anova}

\item{...}{... is the addition argument here}
}
\description{
Printing DF and SS
}
\examples{
library(faraway)
data(coagulation)
attach(coagulation)
k <- onewayanova:::oneway.default(coagulation)
onewayanova:::print.oneway(k)
}
\seealso{
\code{\link{print.oneway}}

\code{\link{print.summary.oneway}}
}

