% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.oneway.R
\name{print.oneway}
\alias{print.oneway}
\title{Print a oneway object}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{an object of class \code{oneway}}

\item{...}{parameters passed to print function}
}
\description{
Prints the results of a oneway ANOVA
}
\examples{
mileage <- oneway(hwy ~ class, cars)
print(mileage)
}
