% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ADehghanBanadaki_HW4_STAT523.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summary.oneway gets input from the oneway.defual to for a summary out of it by making a table}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{object}{object is the ouput of oneway.default which has the values of anova table}

\item{...}{is an addition argument}
}
\description{
Summary.oneway gets input from the oneway.defual to for a summary out of it by making a table
}

