% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.oneway.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summary of One Way ANOVA}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{object of class \code{oneway}}

\item{...}{Parameters passed to print function}
}
\description{
Prints summary of oneway ANOVA
}
